+++ app/vmselect/netstorage/netstorage.go
@@ -453,6 +453,26 @@ func DeleteSeries(sq *storage.SearchQuery) (int, error) {
 	return vmstorage.DeleteMetrics(tfss)
 }
 
+// GetLabelsOnTimeRange returns labels for the given tr until the given deadline.
+func GetLabelsOnTimeRange(tr storage.TimeRange, deadline searchutils.Deadline) ([]string, error) {
+	if deadline.Exceeded() {
+		return nil, fmt.Errorf("timeout exceeded before starting the query processing: %s", deadline.String())
+	}
+	labels, err := vmstorage.SearchTagKeysOnTimeRange(tr, *maxTagKeysPerSearch, deadline.Deadline())
+	if err != nil {
+		return nil, fmt.Errorf("error during labels search on time range: %w", err)
+	}
+	// Substitute "" with "__name__"
+	for i := range labels {
+		if labels[i] == "" {
+			labels[i] = "__name__"
+		}
+	}
+	// Sort labels like Prometheus does
+	sort.Strings(labels)
+	return labels, nil
+}
+
 // GetLabels returns labels until the given deadline.
 func GetLabels(deadline searchutils.Deadline) ([]string, error) {
 	if deadline.Exceeded() {
@@ -462,18 +482,34 @@ func GetLabels(deadline searchutils.Deadline) ([]string, error) {
 	if err != nil {
 		return nil, fmt.Errorf("error during labels search: %w", err)
 	}
-
 	// Substitute "" with "__name__"
 	for i := range labels {
 		if labels[i] == "" {
 			labels[i] = "__name__"
 		}
 	}
-
 	// Sort labels like Prometheus does
 	sort.Strings(labels)
-
 	return labels, nil
+}
+
+// GetLabelValuesOnTimeRange returns label values for the given labelName on the given tr
+// until the given deadline.
+func GetLabelValuesOnTimeRange(labelName string, tr storage.TimeRange, deadline searchutils.Deadline) ([]string, error) {
+	if deadline.Exceeded() {
+		return nil, fmt.Errorf("timeout exceeded before starting the query processing: %s", deadline.String())
+	}
+	if labelName == "__name__" {
+		labelName = ""
+	}
+	// Search for tag values
+	labelValues, err := vmstorage.SearchTagValuesOnTimeRange([]byte(labelName), tr, *maxTagValuesPerSearch, deadline.Deadline())
+	if err != nil {
+		return nil, fmt.Errorf("error during label values search on time range for labelName=%q: %w", labelName, err)
+	}
+	// Sort labelValues like Prometheus does
+	sort.Strings(labelValues)
+	return labelValues, nil
 }
 
 // GetLabelValues returns label values for the given labelName
@@ -485,16 +521,13 @@ func GetLabelValues(labelName string, deadline searchutils.Deadline) ([]string, 
 	if labelName == "__name__" {
 		labelName = ""
 	}
-
 	// Search for tag values
 	labelValues, err := vmstorage.SearchTagValues([]byte(labelName), *maxTagValuesPerSearch, deadline.Deadline())
 	if err != nil {
 		return nil, fmt.Errorf("error during label values search for labelName=%q: %w", labelName, err)
 	}
-
 	// Sort labelValues like Prometheus does
 	sort.Strings(labelValues)
-
 	return labelValues, nil
 }