+++ app/vmselect/prometheus/prometheus.go
@@ -511,9 +511,26 @@ func LabelValuesHandler(startTime time.Time, labelName string, w http.ResponseWr
 		return fmt.Errorf("cannot parse form values: %w", err)
 	}
 	var labelValues []string
-	if len(r.Form["match[]"]) == 0 && len(r.Form["start"]) == 0 && len(r.Form["end"]) == 0 {
+	if len(r.Form["match[]"]) == 0 {
 		var err error
-		labelValues, err = netstorage.GetLabelValues(labelName, deadline)
+		if len(r.Form["start"]) == 0 && len(r.Form["end"]) == 0 {
+			labelValues, err = netstorage.GetLabelValues(labelName, deadline)
+		} else {
+			ct := startTime.UnixNano() / 1e6
+			end, err := searchutils.GetTime(r, "end", ct)
+			if err != nil {
+				return err
+			}
+			start, err := searchutils.GetTime(r, "start", end-defaultStep)
+			if err != nil {
+				return err
+			}
+			tr := storage.TimeRange{
+				MinTimestamp: start,
+				MaxTimestamp: end,
+			}
+			labelValues, err = netstorage.GetLabelValuesOnTimeRange(labelName, tr, deadline)
+		}
 		if err != nil {
 			return fmt.Errorf(`cannot obtain label values for %q: %w`, labelName, err)
 		}
@@ -692,9 +709,26 @@ func LabelsHandler(startTime time.Time, w http.ResponseWriter, r *http.Request) 
 		return fmt.Errorf("cannot parse form values: %w", err)
 	}
 	var labels []string
-	if len(r.Form["match[]"]) == 0 && len(r.Form["start"]) == 0 && len(r.Form["end"]) == 0 {
+	if len(r.Form["match[]"]) == 0 {
 		var err error
-		labels, err = netstorage.GetLabels(deadline)
+		if len(r.Form["start"]) == 0 && len(r.Form["end"]) == 0 {
+			labels, err = netstorage.GetLabels(deadline)
+		} else {
+			ct := startTime.UnixNano() / 1e6
+			end, err := searchutils.GetTime(r, "end", ct)
+			if err != nil {
+				return err
+			}
+			start, err := searchutils.GetTime(r, "start", end-defaultStep)
+			if err != nil {
+				return err
+			}
+			tr := storage.TimeRange{
+				MinTimestamp: start,
+				MaxTimestamp: end,
+			}
+			labels, err = netstorage.GetLabelsOnTimeRange(tr, deadline)
+		}
 		if err != nil {
 			return fmt.Errorf("cannot obtain labels: %w", err)
 		}